+++ packages/users/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @happyvertical/smrt-users
 
+## 0.29.31
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.29.31
+  - @happyvertical/smrt-profiles@0.29.31
+  - @happyvertical/smrt-tenancy@0.29.31
+  - @happyvertical/smrt-config@0.29.31
+  - @happyvertical/smrt-types@0.29.31
+
 ## 0.29.30
 
 ### Patch Changes